Refactor zng_view::config to split platforms into files.
